batch: Extract hunk line range scanning
Live-hunk ownership translation now owns selected-line ownership assembly,
replacement coupling, absence runs, and the scanner primitives that locate
old and new line ranges inside a rendered hunk.

That makes range scanning harder to reason about independently from
ownership mutation, and it leaves hunk_ownership_translation as a dense
coordination point inside the batch layer.

This commit moves the hunk line range scan record and iterators into
hunk_line_ranges, then routes the translator through a module alias so
ownership assembly remains the only responsibility in the original module.

Subsequent commits will add direct scanner coverage and an architecture
guard for the new module boundary.
